ml-service/requirements.txt
# Core ML Service Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLite built-in, SQLAlchemy ORM)
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.9

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pdfplumber>=0.11.0

# Machine Learning
scikit-learn>=1.3.0

# API Requests (for investment advisor)
requests>=2.31.0
